# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Master Plan IT first-party API
  version: 1.0.0
  description: >-
    Static contract for the Laravel API-only backend. The browser uses relative
    same-origin paths through the public React proxy. Laravel is private behind
    that proxy; its internal origin is never sent to browser JavaScript. This
    is not a public developer API.
...
servers:
  - url: /
    description: Same-origin frontend proxy (Laravel remains private)
...
paths:
...
  /api/v1/expenses:
    get:
      tags: [Expenses]
      operationId: listExpenses
      security: [{sessionAuth: []}]
      x-ability: expense.view
      parameters: [{$ref: '#/components/parameters/ExpenseYearQuery'}, {$ref: '#/components/parameters/ExpenseKindQuery'}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PerPage'}]
      responses: {'500': {$ref: '#/components/responses/Unexpected'}, '200': {$ref: '#/components/responses/Expenses'}, '401': {$ref: '#/components/responses/Unauthenticated'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '422': {$ref: '#/components/responses/ValidationError'}}
    post:
      tags: [Expenses]
      operationId: createExpense
      security: [{sessionAuth: []}]
      x-ability: expense.create
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExpenseWrite'}}}}
      responses: {'500': {$ref: '#/components/responses/Unexpected'}, '200': {$ref: '#/components/responses/Expense'}, '401': {$ref: '#/components/responses/Unauthenticated'}, '403': {$ref: '#/components/responses/Forbidden'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /api/v1/expenses/{expense}:
    parameters: [{$ref: '#/components/parameters/ExpenseId'}]
    get:
      tags: [Expenses]
      operationId: showExpense
      security: [{sessionAuth: []}]
      x-ability: expense.view
      responses: {'500': {$ref: '#/components/responses/Unexpected'}, '200': {$ref: '#/components/responses/Expense'}, '401': {$ref: '#/components/responses/Unauthenticated'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}}
    put:
      tags: [Expenses]
      operationId: updateExpense
      security: [{sessionAuth: []}]
      x-ability: expense.update
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExpenseUpdate'}}}}
      responses: {'500': {$ref: '#/components/responses/Unexpected'}, '200': {$ref: '#/components/responses/Expense'}, '401': {$ref: '#/components/responses/Unauthenticated'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
    delete:
      tags: [Expenses]
      operationId: deleteExpense
      security: [{sessionAuth: []}]
      x-ability: expense.delete
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DeleteRequest'}}}}
      responses: {'500': {$ref: '#/components/responses/Unexpected'}, '204': {$ref: '#/components/responses/NoContent'}, '401': {$ref: '#/components/responses/Unauthenticated'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /api/v1/expenses/{expense}/rows/{row}/confirm:
    post:
      tags: [Expenses]
      operationId: confirmActual
      security: [{sessionAuth: []}]
      x-ability: expense.confirm-actual
      parameters: [{$ref: '#/components/parameters/ExpenseId'}, {$ref: '#/components/parameters/RowId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConfirmActualRequest'}}}}
      responses: {'500': {$ref: '#/components/responses/Unexpected'}, '200': {$ref: '#/components/responses/Expense'}, '401': {$ref: '#/components/responses/Unauthenticated'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
...
  /api/v1/contracts/{contract}/generated-expenses/{expense}:
    delete:
      tags: [Contracts]
      operationId: deleteGeneratedExpense
      security: [{sessionAuth: []}]
      x-ability: expense.delete
      parameters: [{$ref: '#/components/parameters/ContractId'}, {$ref: '#/components/parameters/ExpenseId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GeneratedExpenseDelete'}}}}
      responses: {'500': {$ref: '#/components/responses/Unexpected'}, '204': {$ref: '#/components/responses/NoContent'}, '401': {$ref: '#/components/responses/Unauthenticated'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: laravel_session
      description: Laravel Sanctum first-party session cookie; bearer tokens are rejected.
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PerPage: {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 15}}
...
    ExpenseYearQuery: {name: year, in: query, schema: {type: integer, minimum: 1}, description: Identificatore dell'anno di pianificazione.}
    ExpenseKindQuery: {name: kind, in: query, schema: {type: string, enum: [ordinary, plafond]}, description: Natura della spesa; supporta il lookup Plafond dell'editor.}
...
    ExpenseId: {name: expense, in: path, required: true, schema: {type: integer, minimum: 1}}
    RowId: {name: row, in: path, required: true, schema: {type: integer, minimum: 1}}
    ContractId: {name: contract, in: path, required: true, schema: {type: integer, minimum: 1}}
...
  responses:
    NoContent: {description: Mutation completed successfully.}
...
    Expense: {description: Expense detail resource, content: {application/json: {schema: {$ref: '#/components/schemas/ExpenseResponse'}}}}
    Expenses: {description: Paginated expense register, content: {application/json: {schema: {$ref: '#/components/schemas/PaginatedExpenseResponse'}}}}
...
    Unauthenticated: {description: Unauthenticated, headers: {X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Forbidden: {description: Permission or context denied, headers: {X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    NotFound: {description: Not found or protected existence, headers: {X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Conflict: {description: Stale or domain conflict, headers: {X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    ValidationError: {description: Validation failure, headers: {X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
    Unexpected: {description: Unexpected server error, headers: {X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}}, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  headers:
    CorrelationId: {description: Correlation identifier also present in the error payload when available., schema: {type: string}}
  schemas:
...
    LockVersionRequest:
      type: object
      additionalProperties: false
      required: [lock_version]
      properties: {lock_version: {type: integer, minimum: 1}}
...
    DeleteRequest:
      type: object
      additionalProperties: false
      required: [lock_version]
      properties: {lock_version: {type: integer, minimum: 1}, deletion_reason: {type: string, maxLength: 500}}
...
    ExpenseWrite:
      type: object
      additionalProperties: false
      required: [planning_year_id, cost_center_id, kind, title, rows]
      properties:
        planning_year_id: {type: integer, minimum: 1}
        cost_center_id: {type: integer, minimum: 1}
        kind: {type: string}
        title: {type: string, maxLength: 255}
        notes: {type: string}
        contract_id: {type: integer, minimum: 1}
        rows:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/ExpenseRowWrite'}
    ExpenseUpdate:
      allOf: [{$ref: '#/components/schemas/ExpenseWrite'}, {$ref: '#/components/schemas/LockVersionRequest'}, {type: object, properties: {deleted_rows: {type: array, items: {$ref: '#/components/schemas/DeletedRow'}}}}]
    ExpenseRowWrite:
      type: object
      additionalProperties: false
      required: [position, type, description, entered_amount, amount_includes_vat, is_extra]
      properties:
        id: {type: integer, minimum: 1}
        position: {type: integer, minimum: 1}
        vendor_id: {type: integer, minimum: 1}
        type: {type: string}
        description: {type: string, maxLength: 255}
        quantity: {type: string}
        unit_price: {type: string}
        entered_amount: {type: string}
        amount_includes_vat: {type: boolean}
        vat_rate: {type: string}
        is_extra: {type: boolean}
        funded_plafond_expense_id: {type: integer, minimum: 1}
        spend_date: {type: string, format: date}
        period_start: {type: string, format: date}
        period_end: {type: string, format: date}
        distribution: {type: string}
        external_reference: {type: string, maxLength: 255}
        lock_version: {type: integer, minimum: 1}
    DeletedRow:
      type: object
      additionalProperties: false
      required: [id, lock_version]
      properties: {id: {type: integer, minimum: 1}, lock_version: {type: integer, minimum: 1}}
    ConfirmActualRequest:
      type: object
      additionalProperties: false
      required: [lock_version]
      properties: {lock_version: {type: integer, minimum: 1}, confirmed: {type: boolean}}
...
    GeneratedExpenseDelete:
      type: object
      additionalProperties: false
      required: [lock_version, allow_regeneration]
      properties: {lock_version: {type: integer, minimum: 1}, allow_regeneration: {type: boolean}}
    Money:
      type: object
      additionalProperties: false
      required: [net, vat, gross, currency, official_basis]
      properties:
        net: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        vat: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        gross: {type: string, pattern: '^-?[0-9]+(?:\.[0-9]+)?$'}
        currency: {type: string, pattern: '^[A-Z]{3}$'}
        official_basis: {type: string, enum: [net, gross]}
    Meta:
      type: object
      required: [current_page, last_page, per_page, total]
      properties: {current_page: {type: integer}, last_page: {type: integer}, per_page: {type: integer}, total: {type: integer}}
    Links:
      type: object
      required: [first, last, prev, next]
      properties: {first: {type: [string, 'null'], format: uri-reference}, last: {type: [string, 'null'], format: uri-reference}, prev: {type: [string, 'null'], format: uri-reference}, next: {type: [string, 'null'], format: uri-reference}}
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, fields, correlation_id]
          properties: {code: {type: string}, message: {type: string}, fields: {type: object, additionalProperties: true}, correlation_id: {type: [string, 'null']}}
    ErrorResponse: {$ref: '#/components/schemas/Error'}
...
    ExpenseRow:
      type: object
      additionalProperties: false
      properties:
        id: {type: integer}
        position: {type: integer}
        vendor_id: {type: [integer, 'null']}
        vendor_name: {type: [string, 'null']}
        type: {type: string}
        confirmation_state: {type: [string, 'null']}
        description: {type: string}
        quantity: {type: [string, 'null']}
        unit_price: {type: [string, 'null']}
        entered_amount: {type: string}
        amount_includes_vat: {type: boolean}
        vat_rate: {type: [string, 'null']}
        is_extra: {type: boolean}
        funded_plafond_expense_id: {type: [integer, 'null']}
        spend_date: {type: [string, 'null'], format: date}
        period_start: {type: [string, 'null'], format: date}
        period_end: {type: [string, 'null'], format: date}
        distribution: {type: [string, 'null']}
        external_reference: {type: [string, 'null']}
        lock_version: {type: integer}
        is_system_managed: {type: boolean}
        generated: {type: boolean}
        contract_term_id: {type: [integer, 'null']}
        totals: {$ref: '#/components/schemas/Money'}
    Expense:
      type: object
      additionalProperties: false
      properties:
        id: {type: integer}
        planning_year_id: {type: integer}
        planning_year_label: {type: integer}
        cost_center_id: {type: integer}
        cost_center_name: {type: [string, 'null']}
        kind: {type: string}
        title: {type: string}
        notes: {type: [string, 'null']}
        contract_id: {type: [integer, 'null']}
        lock_version: {type: integer}
        rows: {type: array, items: {$ref: '#/components/schemas/ExpenseRow'}}
        totals: {$ref: '#/components/schemas/Money'}
...
    ExpenseResponse: {type: object, required: [data], properties: {data: {$ref: '#/components/schemas/Expense'}}}
    PaginatedExpenseResponse: {type: object, required: [data, meta, links], properties: {data: {type: array, items: {$ref: '#/components/schemas/Expense'}}, meta: {$ref: '#/components/schemas/Meta'}, links: {$ref: '#/components/schemas/Links'}}}
...